Add card service CRUD implementation with validations
Introduced full CRUD functionality for managing cards, including service, controller, repository, and DTOs. Added exception handling and validation mechanisms to ensure robust error reporting and input validation. Updated project structure to include card-related REST APIs and relevant configurations.

+++ src/main/java/com/ritik/cardsservice/dto/CardDTO.java
@@ -34,17 +34,17 @@ public class CardDTO {
     @Schema(
             description = "Total amount limit available against a card", example = "100000"
     )
-    private int totalLimit;
+    private Double totalLimit;
 
     @PositiveOrZero(message = "Total amount used should be equal or greater than zero")
     @Schema(
             description = "Total amount used by a Customer", example = "1000"
     )
-    private int amountUsed;
+    private Double amountUsed;
 
     @PositiveOrZero(message = "Total available amount should be equal or greater than zero")
     @Schema(
             description = "Total available amount against a card", example = "90000"
     )
-    private int availableAmount;
+    private Double availableAmount;
 }
